Use early stopping
* keep track of the best validation accuracy
* stop if validation accuracy doesn't improve for a pre-specified number of
epochs.

Motivation
While the training error may constantly improve with more epochs,
but it might lead to overfitting. In order to prevent this, we measure the
generalization error on the validation set, and stop the training process if
it starts to deteriorate. Early stopping can be viewed as a hyper-parameter.

Signed-off-by: Dimid Duchovny <dimidd@gmail.com>